services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
prometheus-client==0.19.0
slowapi==0.1.9
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1

